[MJCF] Fixed issue with keyframe initialization for free base
When the base is free additional degrees of freedom for the base need to be considered
that are not defined in the sdf/animat morphology. This has now been addressed. There is
an error in MuJoCo documentation where qpos and qvel are said to be of the same
dimensions but infact qpos is nqx1 qvel is nvx1